Fix CI/CD: Add size_t cast to fix Clang sign-conversion warning
Fixed implicit sign conversion in udatamemoffset():
- Cast nuv to size_t before multiplication with sizeof(UValue)
- Eliminates -Wsign-conversion error in Clang with -Werror

The multiplication of sizeof(UValue) * nuv was implicitly converting
int to unsigned long, which Clang treats as an error.

+++ src/objects/lobject.h
@@ -749,7 +749,7 @@ typedef struct Udata0 : public GCBase<Udata0> {
 // This triggers -Winvalid-offsetof but is safe because we control the memory layout
 constexpr inline size_t udatamemoffset(int nuv) noexcept {
 	return (nuv == 0) ? offsetof(Udata0, bindata)
-	                  : Udata::uvOffset() + (sizeof(UValue) * nuv);
+	                  : Udata::uvOffset() + (sizeof(UValue) * static_cast<size_t>(nuv));
 }
 
 /* get the address of the memory block inside 'Udata' */
